// src/payload-types.ts
export interface Tenant {
  id: number;
  name: string;
  domains?:
    | {
        domain: string;
        id?: string | null;
      }[]
    | null;
  /**
   * Used for url paths, example: /tenant-slug/page-slug
   */
  slug: string;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Role {
  id: number;
  name: string;
  rules: {
    collections: string[];
    actions: ('*' | 'create' | 'read' | 'update' | 'delete')[];
    id?: string | null;
  }[];
  updatedAt: string;
  createdAt: string;
}